Extract test class definitions to separate fixture files (#1275)
