+++ tests/Unit/Schema/Directives/Fixtures/Foo.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Tests\Unit\Schema\Directives\Fixtures;
+
+/**
+ * TODO remove in favour of ->getMockBuilder(\stdClass::class)->addMethods(['__invoke'])
+ * once we no longer support PHPUnit 7.
+ */
+class Foo
+{
+    public function bar()
+    {
+        //
+    }
+}
